backend/requirements.txt
fastapi
uvicorn
langchain
openai
reportlab
python-dotenv
